Add Token-based Authorization Option
RFC7662 specifies that you MUST use authentication for an introspection
request, but it doesn't specify that it must be client-based:

"To prevent token scanning attacks, the endpoint MUST also require
some form of authorization to access this endpoint, such as client
authentication as described in OAuth 2.0 [RFC6749] or a separate
OAuth 2.0 access token such as the bearer token described in OAuth
2.0 Bearer Token Usage [RFC6750].  The methods of managing and
validating these authentication credentials are out of scope of this
specification."

Our OAuth server does a BCrypt hash for client-based authorization (and
any OAuth server should do some type of expensive operation on the
client secret) and so it also supports client tokens for an
introspection request.

This PR adds support for these in addition to client ID/secret
authorization.